Make a separate file for the effectful code in the reducer
